ISSUE #3705 - Minor refactor of ticket image and ticketview
