+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsForm/properties/basicTicketImage/ticketImage/ticketImage.component.tsx
@@ -15,38 +15,11 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import AddImageIcon from '@assets/icons/outlined/add_image-outlined.svg';
-import EditImageIcon from '@assets/icons/outlined/edit-outlined.svg';
-import { FormattedMessage } from 'react-intl';
 import { useEffect } from 'react';
-import { ActionMenuItem } from '@controls/actionMenu/actionMenuItem/actionMenuItem.component';
-import { Viewer as ViewerService } from '@/v4/services/viewer/viewer';
-import { addBase64Prefix, convertFileToImageSrc, getSupportedImageExtensions, stripBase64Prefix } from '@controls/fileUploader/imageFile.helper';
-import { uploadFile } from '@controls/fileUploader/uploadFile';
-import { useParams } from 'react-router-dom';
-import { getTicketResourceUrl, isResourceId, modelIsFederation } from '@/v5/store/tickets/tickets.helpers';
-import { TicketsCardHooksSelectors } from '@/v5/services/selectorsHooks';
-import { ActionMenu, MenuItem, MenuItemDelete } from '../ticketImageAction/ticketImageAction.styles';
-import { TicketImageAction } from '../ticketImageAction/ticketImageAction.component';
+import { stripBase64Prefix } from '@controls/fileUploader/imageFile.helper';
+import { getImgSrc } from '@/v5/store/tickets/tickets.helpers';
 import { BasicTicketImage, BasicTicketImageProps } from '../basicTicketImage.component';
-
-const TriggerButton = ({ hasImage }) => {
-	if (!hasImage) {
-		return (
-			<TicketImageAction>
-				<AddImageIcon />
-				<FormattedMessage id="viewer.card.ticketImage.action.addImage" defaultMessage="Add image" />
-			</TicketImageAction>
-		);
-	}
-
-	return (
-		<TicketImageAction>
-			<EditImageIcon />
-			<FormattedMessage id="viewer.card.ticketImage.action.editImage" defaultMessage="Edit image" />
-		</TicketImageAction>
-	);
-};
+import { TicketImageActionMenu } from '../ticketImageActionMenu.component';
 
 type TicketImageProps = Omit<BasicTicketImageProps, 'onEmptyImageClick' | 'imgSrc' | 'children'> & {
 	onChange?: (imgSrc) => void,
@@ -54,54 +27,18 @@ type TicketImageProps = Omit<BasicTicketImageProps, 'onEmptyImageClick' | 'imgSr
 	value?: string,
 };
 export const TicketImage = ({ value, onChange, onBlur, ...props }: TicketImageProps) => {
-	const { teamspace, project, containerOrFederation } = useParams();
-	const ticketId = TicketsCardHooksSelectors.selectSelectedTicketId();
-	const isFederation = modelIsFederation(containerOrFederation);
-	const hasImage = !!value;
-
-	const handleImageChange = (newValue) => onChange(stripBase64Prefix(newValue));
-
-	const uploadScreenshot = async () => handleImageChange(await ViewerService.getScreenshot());
-
-	const uploadImage = async () => {
-		const file = await uploadFile(getSupportedImageExtensions());
-		const imgSrc = await convertFileToImageSrc(file);
-		handleImageChange(imgSrc);
-	};
-
-	const deleteImage = () => handleImageChange('');
-
-	const getImgSrc = (imgData) => {
-		if (!imgData) return '';
-		if (isResourceId(imgData)) {
-			return getTicketResourceUrl(teamspace, project, containerOrFederation, ticketId, value, isFederation);
-		}
-		return addBase64Prefix(imgData);
-	};
+	const onImageChange = (newValue) => onChange(newValue ? stripBase64Prefix(newValue) : null);
+	const imgSrc = getImgSrc(value);
 
 	useEffect(() => { setTimeout(() => { onBlur?.(); }, 200); }, [value]);
 
 	return (
 		<BasicTicketImage
-			imgSrc={getImgSrc(value)}
-			onEmptyImageClick={uploadImage}
+			value={imgSrc}
+			onChange={onImageChange}
 			{...props}
 		>
-			<ActionMenu TriggerButton={<div><TriggerButton hasImage={hasImage} /></div>}>
-				<ActionMenuItem>
-					<MenuItem onClick={uploadScreenshot}>
-						<FormattedMessage id="viewer.card.ticketImage.action.createScreenshot" defaultMessage="Create screenshot" />
-					</MenuItem>
-					<MenuItem onClick={uploadImage}>
-						<FormattedMessage id="viewer.card.ticketImage.action.uploadImage" defaultMessage="Upload image" />
-					</MenuItem>
-					{hasImage && (
-						<MenuItemDelete onClick={deleteImage}>
-							<FormattedMessage id="viewer.card.ticketImage.action.deleteImage" defaultMessage="Delete image" />
-						</MenuItemDelete>
-					)}
-				</ActionMenuItem>
-			</ActionMenu>
+			<TicketImageActionMenu value={imgSrc} onChange={onImageChange} />
 		</BasicTicketImage>
 	);
 };
