+++ frontend/src/v5/ui/routes/viewer/tickets/ticketsForm/properties/basicTicketImage/ticketImageActionMenu.component.tsx
@@ -44,7 +44,7 @@ const TriggerButton = ({ hasImage }) => {
 	);
 };
 
-export const TicketImageActionButton = ({ value, onChange }) => {
+export const TicketImageActionMenu = ({ value, onChange }) => {
 	const uploadScreenshot = async () => onChange(await ViewerService.getScreenshot());
 
 	const uploadImage = async () => {
